requirements.txt
certifi==2017.4.17
chardet==3.0.4
Django<=1.8
django-extensions==1.7.9
django-upwork-auth
httplib2==0.10.3
httplib2.system-ca-certs-locater==0.1.2
idna==2.5
oauth2==1.5.211
pkg-resources==0.0.0
pprint==0.1
pydot==1.2.3
pyparsing==2.2.0
python-upwork>=1.0
pytz==2017.2
requests==2.18.1
six==1.10.0
urllib3==1.21.1
